fix: restore StMakeEnvelope to GeoExpressions.scala to fix scalariform parse error
Scalariform 0.2.6's stateful parser requires StMakeEnvelope to be present
between StPoint and StMakeLine. Without it the parser fails at the first =>
after StPoint. Deleted the now-redundant StMakeEnvelopeExpr.scala.
